Fix shard ranges in seedtos3.sh
The shard_seeder tool (in opencbdc-tx) computes shard range
as 256/#shards, i.e., rounds down the range. This script computes
it as 255/#shards +1, i.e., rounds it up. This doesn't matter if
num shards is a power of 2, but for other values, the tool and script
differ, causing errors in seed generation. The quick fix in this
commit is to calculate the range by rounding down, as the tool
does, and to ensure that the last range is extended, if needed,
to the 1-byte max or 255.

Signed-off-by: Ram Darbha <darbha.ram@gmail.com>